Do evaluation of tag in unwrap in constant time
The time needed to check the tag in spongewrap_unwrap() should now only
depend on the length of the tag and not on its correctness.

I'm not sure whether or not this is necessary but I don't see how it
could hurt either.

Before this commit, an attacker could determine via the timing, which
byte of the tag was invalid. However, since the spongewrap's state
changes during the check, brute forcing the tag byte by byte should be
impossible.

The only way I can think of would be if spongewrap is used in some form
of authentication, where the structure is destroyed and then recreated
after each failed attempt.