Reduce the use of explicit_bzero
Profiling showed the excessive use of explicit_bzero to be very
expensive. I replaced calls to explicit_bzero with calls to memset in
the padding and in the set_nonce function. It should not cause any
trouble there.
I also removed the bzero calls in duplex_duplexing and
duplex_with_frame_bit. Both buffers are now cleared once spongewrap_wrap
or spongewrap_unwrap returns.
Lastly, I shrunk the buffer in mcfd_net. It was vastly oversized and
explicit_bzero was called on it after every forwarded packet.